+++ internal/ipc/client.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/ActiveState/cli/exp/pm/internal/ipc/internal/flisten"
+	"github.com/ActiveState/cli/internal/ipc/internal/flisten"
 )
 
 type Client struct {
