+++ internal/ipc/comm.go
