+++ internal/ipc/namespace/namespace.go
