+++ internal/svccomm/svccomm.go
@@ -4,7 +4,7 @@ package svccomm
 import (
 	"context"
 
-	"github.com/ActiveState/cli/exp/pm/internal/ipc"
+	"github.com/ActiveState/cli/internal/ipc"
 )
 
 var (
