+++ test/pseudo/cmd/tool/main.go
@@ -7,9 +7,9 @@ import (
 	"path/filepath"
 	"time"
 
-	"github.com/ActiveState/cli/exp/pm/cmd/internal/serve"
-	"github.com/ActiveState/cli/exp/pm/internal/ipc"
-	"github.com/ActiveState/cli/exp/pm/internal/svcctl"
+	"github.com/ActiveState/cli/internal/ipc"
+	"github.com/ActiveState/cli/internal/svcctl"
+	"github.com/ActiveState/cli/test/pseudo/cmd/internal/serve"
 )
 
 func main() {
